feat(primitives): Move deal structs from market to primitives
