+++ pallets/market/src/lib.rs
@@ -22,30 +22,28 @@ pub mod pallet {
         ensure,
         pallet_prelude::*,
         sp_runtime::{
-            traits::{
-                AccountIdConversion, CheckedAdd, CheckedSub, Hash, IdentifyAccount, Verify, Zero,
-            },
+            traits::{AccountIdConversion, CheckedAdd, CheckedSub, Hash, Verify, Zero},
             ArithmeticError, BoundedBTreeMap, RuntimeDebug,
         },
         traits::{
             Currency,
             ExistenceRequirement::{AllowDeath, KeepAlive},
-            Hooks, ReservableCurrency, WithdrawReasons,
+            Hooks, WithdrawReasons,
         },
         PalletId,
     };
-    use frame_system::{pallet_prelude::*, Config as SystemConfig, Pallet as System};
+    use frame_system::{pallet_prelude::*, Pallet as System};
     use primitives::{
         commitment::{
             commd::compute_unsealed_sector_commitment,
             piece::{PaddedPieceSize, PieceInfo},
-            CommP, Commitment, CommitmentError,
         },
-        configs::{CurrencyProvider, MarketProvider},
+        configs::{BalanceOf, CurrencyProvider, MarketProvider},
+        deals::{ActiveDealState, ClientDealProposal, DealProposal, DealProposalOf, DealState},
         pallets::{ActiveDeal, ActiveSector, Market, SectorDeal, StorageProviderValidation},
         proofs::RegisteredSealProof,
         sector::{SectorNumber, SectorSize},
-        DealId, CID_SIZE_IN_BYTES, MAX_DEALS_PER_SECTOR, MAX_LABEL_SIZE, MAX_SECTORS_PER_CALL,
+        DealId, MAX_DEALS_PER_SECTOR, MAX_SECTORS_PER_CALL,
     };
     use scale_info::TypeInfo;
     use sp_arithmetic::traits::BaseArithmetic;
@@ -102,173 +100,6 @@ pub mod pallet {
         /// Amount of Balance that has been staked as Deal Collateral
         /// It's locked to a deal and cannot be withdrawn until the deal ends.
         pub locked: Balance,
-    }
-
-    #[derive(Clone, Eq, PartialEq, Encode, Decode, RuntimeDebug, TypeInfo, MaxEncodedLen)]
-    pub enum DealState<BlockNumber> {
-        /// Deal has been accepted on-chain by both Storage Provider and Storage Client, it's waiting for activation.
-        Published,
-        /// Deal has been activated
-        Active(ActiveDealState<BlockNumber>),
-    }
-
-    #[derive(Clone, Eq, PartialEq, Encode, Decode, RuntimeDebug, TypeInfo, MaxEncodedLen)]
-    /// State only related to the activated deal
-    /// Reference: <https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/deal.rs#L138>
-    pub struct ActiveDealState<BlockNumber> {
-        /// Sector in which given piece has been included
-        pub sector_number: SectorNumber,
-
-        /// At which block (time) the deal's sector has been activated.
-        pub sector_start_block: BlockNumber,
-
-        /// The last block (time) when the deal was updated — i.e. when a deal payment settlement was made.
-        ///
-        /// In Filecoin this happens under two circumstances:
-        /// * Someone starts the payment settlement procedure.
-        /// * Cron tick (deprecated) settles legacy deals.
-        ///
-        /// Sources:
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/lib.rs#L985>
-        /// * <https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/lib.rs#L1315>
-        pub last_updated_block: Option<BlockNumber>,
-
-        /// When the deal was last slashed, can be never.
-        ///
-        /// In Filecoin, slashing can happen in two cases, storage faults and consensus faults,
-        /// in our case, we're only concerned about the storage faults, as the consensus is
-        /// handled by the collators.
-        ///
-        /// Slashing is related to three main kinds of penalties:
-        /// * Fault Fee — incurred for each day a sector is offline.
-        /// * Storage Penalty — incurred when sectors that were not declared as faulty before a WindowPoSt are detected.
-        /// * Termination Penalty — incurred when a sector is voluntarily (the miner "gave up on the deal") or
-        ///   involuntarily (when a sector is faulty for 42 days in a row) terminated and removed from the network.
-        ///
-        /// Slashing is applied (i.e. `slash_epoch` is updated) in a single place:
-        /// * During [`on_miners_sector_terminate`][1], by termination penalty since the deal was terminated early.
-        ///   The deal is first settled — i.e. the storage provider gets paid for the storage time since they last settled the deal —
-        ///   then storage provider has their collateral slashed and burned and the client gets their funds unlocked (i.e. refunded).
-        ///
-        /// However, slashing is performed in other places, it just does not update `slash_epoch` (`slash_block` in our case).
-        /// * During [`get_active_deal_or_process_timeout`][2], slashing will happen if the deal has expired
-        ///   — i.e. if and when the deal is published but fails to be activated in a given period.
-        ///   This function is called in [`cron_tick`][3] and [`settle_deal_payments`][4].
-        /// * During [`process_deal_update`][5], if the deal has a `slash_epoch`, any remaining payments will be settled
-        ///   and the provider will have its collateral slashed.
-        /// * During [`cron_tick`][7], by means of [`get_active_deal_or_process_timeout`][8] and finally [`process_deal_init_timed_out`][9].
-        ///
-        /// Sources:
-        /// * <https://spec.filecoin.io/#section-glossary.storage-fault-slashing>
-        /// * <https://spec.filecoin.io/#section-systems.filecoin_mining.sector.lifecycle>
-        ///
-        /// [1]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/lib.rs#L852-L853
-        /// [2]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/state.rs#L741-L797
-        /// [3]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/lib.rs#L904-L924
-        /// [4]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/lib.rs#L1240-L1271
-        /// [5]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/state.rs#L886-L912
-        /// [6]: https://github.com/filecoin-project/builtin-actors/blob/54236ae89880bf4aa89b0dba6d9060c3fd2aacee/actors/market/src/state.rs#L922-L962
-        /// [7]: https://github.com/filecoin-project/builtin-actors/blob/54236ae89880bf4aa89b0dba6d9060c3fd2aacee/actors/market/src/lib.rs#L904-L924
-        /// [8]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/state.rs#L765
-        /// [9]: https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/state.rs#L964-L997
-        pub slash_block: Option<BlockNumber>,
-    }
-
-    impl<BlockNumber> ActiveDealState<BlockNumber> {
-        pub fn new(
-            sector_number: SectorNumber,
-            sector_start_block: BlockNumber,
-        ) -> ActiveDealState<BlockNumber> {
-            ActiveDealState {
-                sector_number,
-                sector_start_block,
-                last_updated_block: None,
-                slash_block: None,
-            }
-        }
-    }
-
-    #[derive(Clone, Eq, PartialEq, Encode, Decode, RuntimeDebug, TypeInfo, MaxEncodedLen)]
-    /// Reference: <https://github.com/filecoin-project/builtin-actors/blob/17ede2b256bc819dc309edf38e031e246a516486/actors/market/src/deal.rs#L93>
-    // It cannot be generic over <T: Config> because, #[derive(RuntimeDebug, TypeInfo)] also make `T` to have `RuntimeDebug`/`TypeInfo`
-    // It is a known rust issue <https://substrate.stackexchange.com/questions/452/t-doesnt-implement-stdfmtdebug>
-    pub struct DealProposal<Address, Balance, BlockNumber> {
-        /// Byte Encoded Cid
-        // We use BoundedVec here, as cid::Cid do not implement `TypeInfo`, so it cannot be saved into the Runtime Storage.
-        // It maybe doable using newtype pattern, however not sure how the UI on the frontend side would handle that anyways.
-        // There is Encode/Decode implementation though, through the feature flag: `scale-codec`.
-        pub piece_cid: BoundedVec<u8, ConstU32<CID_SIZE_IN_BYTES>>,
-        /// The value represents the size of the data piece after padding to the
-        /// nearest power of two. Padding ensures that all pieces can be
-        /// efficiently arranged in a binary tree structure for Merkle proofs.
-        pub piece_size: u64,
-        /// Storage Client's Account Id
-        pub client: Address,
-        /// Storage Provider's Account Id
-        pub provider: Address,
-
-        /// Arbitrary client chosen label to apply to the deal
-        pub label: BoundedVec<u8, ConstU32<MAX_LABEL_SIZE>>,
-
-        /// Nominal start block. Deal payment is linear between StartBlock and EndBlock,
-        /// with total amount StoragePricePerBlock * (EndBlock - StartBlock).
-        /// Storage deal must appear in a sealed (proven) sector no later than StartBlock,
-        /// otherwise it is invalid.
-        pub start_block: BlockNumber,
-        /// When the Deal is supposed to end.
-        pub end_block: BlockNumber,
-        /// `Deal` can be terminated early, by `on_sectors_terminate`.
-        /// Before that, a Storage Provider can payout it's earned fees by calling `on_settle_deal_payments`.
-        /// `on_settle_deal_payments` must know how much money it can payout, so it's related to the number of blocks (time) it was stored.
-        /// Reference <https://spec.filecoin.io/#section-systems.filecoin_markets.onchain_storage_market.storage_deal_states>
-        pub storage_price_per_block: Balance,
-
-        /// Amount of Balance (DOTs) Storage Provider stakes as Collateral for storing given `piece_cid`
-        /// There should be enough Balance added by `add_balance` by Storage Provider to cover it.
-        /// When the Deal fails/is terminated to early, this is the amount which get slashed.
-        pub provider_collateral: Balance,
-        /// Current [`DealState`].
-        /// It goes: `Published` -> `Active`
-        pub state: DealState<BlockNumber>,
-    }
-
-    impl<Address, Balance, BlockNumber> DealProposal<Address, Balance, BlockNumber>
-    where
-        Balance: BaseArithmetic + Copy,
-        BlockNumber: BaseArithmetic + Copy,
-    {
-        fn duration(&self) -> BlockNumber {
-            self.end_block - self.start_block
-        }
-
-        fn total_storage_fee(&self) -> Option<u128> {
-            // We need to convert into something to perform the calculation.
-            // Generics trickery prevents us from doing it in a nice way.
-            // <https://stackoverflow.com/questions/56081117/how-do-you-convert-between-substrate-specific-types-and-rust-primitive-types>
-            Some(
-                TryInto::<u128>::try_into(self.storage_price_per_block).ok()?
-                    * TryInto::<u128>::try_into(self.duration()).ok()?,
-            )
-        }
-
-        pub fn piece_commitment(&self) -> Result<Commitment<CommP>, CommitmentError> {
-            let commitment = Commitment::from_cid_bytes(&self.piece_cid[..])?;
-            Ok(commitment)
-        }
-    }
-
-    type DealProposalOf<T> =
-        DealProposal<<T as frame_system::Config>::AccountId, BalanceOf<T>, BlockNumberFor<T>>;
-
-    /// After Storage Client has successfully negotiated with the Storage Provider, they prepare a DealProposal,
-    /// sign it with their signature and send to the Storage Provider.
-    /// Storage Provider only after successful file transfer and verification of the data, calls an extrinsic `market.publish_storage_deals`.
-    /// The extrinsic call is signed by the Storage Provider and Storage Client's signature is in the message.
-    /// Based on that, Market Pallet can verify the signature and lock appropriate funds.
-    #[derive(Clone, Eq, PartialEq, Encode, Decode, RuntimeDebug, TypeInfo, MaxEncodedLen)]
-    pub struct ClientDealProposal<Address, Currency, BlockNumber, OffchainSignature> {
-        pub proposal: DealProposal<Address, Currency, BlockNumber>,
-        pub client_signature: OffchainSignature,
     }
 
     #[pallet::pallet]
