+++ primitives/Cargo.toml
@@ -21,6 +21,7 @@ sealed = { workspace = true }
 serde = { workspace = true, optional = true, features = ["derive"] }
 sha2 = { workspace = true }
 sp-api = { workspace = true }
+sp-arithmetic = { workspace = true, default-features = false }
 sp-core = { workspace = true }
 sp-io = { workspace = true, default-features = false }
 sp-runtime = { workspace = true }
