+++ primitives/src/deals/mod.rs
@@ -1,0 +1,9 @@
+mod active_deal_state;
+mod client_deal_proposal;
+mod deal_proposal;
+mod deal_state;
+
+pub use active_deal_state::*;
+pub use client_deal_proposal::*;
+pub use deal_proposal::*;
+pub use deal_state::*;
